Change the name of the brick and brick parts
The brick and brickpart is a smell as the name dont tell what the do. Brick is a containers of Brickparts that is just struct for one Brickparts.
I change the names from Brick to BrickContainer as it tells what it is and BrickParts to Brick as it just one brick.